pyproject.toml
[tool.poetry]
name = "botfactory"
version = "0.1.0"
description = "AI Chatbot Platform"
authors = ["AI Developer"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "3.0.0"
SQLAlchemy = "2.0.29"
Flask-SQLAlchemy = "3.1.1"
Flask-Login = "0.6.3"
Flask-Migrate = "4.0.7"
Flask-Babel = "4.0.0"
google-generativeai = "0.5.4"
python-telegram-bot = {version = "21.2", extras = ["ext"]}
gunicorn = "22.0.0"
psycopg2-binary = "*"
werkzeug = "3.0.1"
